Backend/requirements.txt
fastapi
tortoise-orm
uvicorn
aiomysql
passlib
python-jose